Renamed apiRoutes.js and Roles.js, customized index.js, reformatted server.js
