+++ models/Roles.js
@@ -1,0 +1,3 @@
+// models are static files. how we interface with our ORM, Sequelize, and our database
+// API is separate from single-page web app/static pages. Both are being served by server
+// views are the name of HTML pages that you build for your app
