+++ models/index.js
@@ -1,11 +1,25 @@
-import DietaryRestrictions from './DietaryRestrictions.js';
-import DiningHall from './DiningHall.js';
-import Macros from './Macros.js';
-import Meals from './Meals.js';
+import Media from './Media.js';
+import Creators from './DiningHall.js';
+import Backgrounds from './Macros.js';
+import Genres from './Meals.js';
+import Roles from './Meals.js';
+import Themes from './Meals.js';
+import MediaBackgroundLinks from './Meals.js';
+import   MediaCreatorLinks from './Meals.js';
+import MediaGenreLinks from './Meals.js';
+import MediaThemeLinks from './Meals.js';
+import CreatorRoleLinks from './Meals.js';
 
 export default {
-  DietaryRestrictions,
-  DiningHall,
-  Macros,
-  Meals
+  Media,
+  Creators,
+  Backgrounds,
+  Genres,
+  Roles,
+  Themes,
+  MediaBackgroundLinks,
+  MediaCreatorLinks,
+  MediaGenreLinks,
+  MediaThemeLinks,
+  CreatorRoleLinks,
 };
