+++ routes/apiRoutes.js
@@ -1,4 +1,6 @@
 /* eslint-disable no-console */
+
+// export into controllers later?
 import express from "express";
 import sequelize from "sequelize";
 
@@ -7,17 +9,388 @@ import db from "../database/initializeDB.js";
 const router = express.Router();
 
 router.get("/", (req, res) => {
-  res.send("Welcome to the UMD Dining API!");
+  res.send("Welcome to the AAPI Art Corner API!");
 });
 
 /// /////////////////////////////////
-/// ////Dining Hall Endpoints////////
+/// ////Media Endpoints////////
 /// /////////////////////////////////
-router.get("/dining", async (req, res) => {
+router
+  .route("/Media")
+  .get(async (req, res) => {
+    try {
+      console.log("GET request on /Media");
+      const media = await db.Media.findAll();
+      const reply =
+        media.length > 0 ? { data: media } : { message: "no results found" };
+      res.json(reply);
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  })
+  .post(async (req, res) => {
+    const media = await db.AllMedia.findAll();
+    const currentId = (await media.length) + 1;
+    try {
+      console.log("POST request on /media");
+      // use CamelCase (example Media)
+      const newMedia = await db.all_media.create({
+        media_id: currentId,
+        media_title: req.body.media_title,
+        media_type: req.body.media_type,
+        media_release_year: req.body.media_release_year,
+        media_description: req.body.media_description,
+        media_duration: req.body.media_duration,
+        album_songs_number: req.body.album_songs_number,
+        television_seasons_number: req.body.television_seasons_number,
+        audience_rating: req.body.audience_rating,
+        show_still_airing: req.body.show_still_airing
+      });
+      res.json(newMedia);
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  })
+  .put(async (req, res) => {
+    try {
+      console.log("PUT request on /media");
+      await db.all_media.update(
+        {
+          media_title: req.body.media_title,
+          media_type: req.body.media_type,
+          media_release_year: req.body.media_release_year,
+          media_description: req.body.media_description,
+          media_duration: req.body.media_duration,
+          album_songs_number: req.body.album_songs_number,
+          television_seasons_number: req.body.television_seasons_number,
+          audience_rating: req.body.audience_rating,
+          show_still_airing: req.body.show_still_airing
+        },
+        {
+          where: {
+            media_id: req.body.media_id
+          }
+        }
+      );
+      res.send("Successfully Updated");
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  })
+  .delete((req, res) => {
+    try {
+      console.log("DELETE request on /media");
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  });
+
+/// /////////////////////////////////
+/// ////Creators Endpoints////////
+/// /////////////////////////////////
+
+router
+  .route("/creators")
+  .get(async (req, res) => {
+    try {
+      console.log("GET request on /creators");
+      const creators = await db.creators.findAll();
+      const reply =
+        creators.length > 0
+          ? { data: creators }
+          : { message: "no results found" };
+      res.json(reply);
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  })
+  .post(async (req, res) => {
+    const creators = await db.creators.findAll();
+    const currentId = (await creators.length) + 1;
+    try {
+      console.log("POST request on /creators");
+      const newCreator = await db.creators.create({
+        creator_id: currentId,
+        creator_first_name: req.body.creator_first_name,
+        creator_last_name: req.body.creator_last_name,
+        creator_current_state: req.body.creator_current_state,
+        creator_home_state: req.body.creator_home_state,
+        creator_country: req.body.creator_country
+      });
+      res.json(newCreator);
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  })
+  .put(async (req, res) => {
+    try {
+      console.log("PUT request on /creators");
+      await db.creators.update(
+        {
+          creator_first_name: req.body.creator_first_name,
+          creator_last_name: req.body.creator_last_name,
+          creator_current_state: req.body.creator_current_state,
+          creator_home_state: req.body.creator_home_state,
+          creator_country: req.body.creator_country
+        },
+        {
+          where: {
+            creator_id: req.body.creator_id
+          }
+        }
+      );
+      res.send("Successfully Updated");
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  })
+  .delete((req, res) => {
+    res.send("Action unavailable");
+  });
+
+/// /////////////////////////////////
+/// //// Backgrounds Endpoints////////
+/// /////////////////////////////////
+
+router
+  .route("/backgrounds")
+  .get(async (req, res) => {
+    try {
+      console.log("GET request on /backgrounds");
+      const backgrounds = await db.backgrounds.findAll();
+      const reply =
+        backgrounds.length > 0
+          ? { data: backgrounds }
+          : { message: "no results found" };
+      res.json(reply);
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  })
+  .post(async (req, res) => {
+    const backgrounds = await db.backgrounds.findAll();
+    const currentId = (await backgrounds.length) + 1;
+    try {
+      console.log("POST request on /backgrounds");
+      const newBackground = await db.backgrounds.create({
+        background_id: currentId,
+        background: req.body.background
+      });
+      res.json(newBackground);
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  })
+  .put(async (req, res) => {
+    try {
+      console.log("PUT request on /backgrounds");
+      await db.backgrounds.update(
+        {
+          background: req.body.background
+        },
+        {
+          where: {
+            background_id: req.body.background_id
+          }
+        }
+      );
+      res.send("Successfully Updated");
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  })
+  .delete((req, res) => {
+    res.send("Action unavailable");
+  });
+
+/// /////////////////////////////////
+/// //// Genres Endpoints////////
+/// /////////////////////////////////
+
+router
+  .route("/genres")
+  .get(async (req, res) => {
+    try {
+      console.log("GET request on /genres");
+      const genres = await db.genres.findAll();
+      const reply =
+        genres.length > 0 ? { data: genres } : { message: "no results found" };
+      res.json(reply);
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  })
+  .post(async (req, res) => {
+    const genres = await db.genres.findAll();
+    const currentId = (await genres.length) + 1;
+    try {
+      console.log("POST request on /genres");
+      const newGenre = await db.genres.create({
+        genre_id: currentId,
+        genre: req.body.genre
+      });
+      res.json(newGenre);
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  })
+  .put(async (req, res) => {
+    try {
+      console.log("PUT request on /genres");
+      await db.genres.update(
+        {
+          genre: req.body.genre
+        },
+        {
+          where: {
+            genre_id: req.body.genre_id
+          }
+        }
+      );
+      res.send("Successfully Updated");
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  })
+  .delete((req, res) => {
+    res.send("Action unavailable");
+  });
+
+/// /////////////////////////////////
+/// //// Roles Endpoints////////
+/// /////////////////////////////////
+
+router
+  .route("/roles")
+  .get(async (req, res) => {
+    try {
+      console.log("GET request on /roles");
+      const roles = await db.roles.findAll();
+      const reply =
+        roles.length > 0 ? { data: roles } : { message: "no results found" };
+      res.json(reply);
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  })
+  .post(async (req, res) => {
+    const roles = await db.roles.findAll();
+    const currentId = (await roles.length) + 1;
+    try {
+      console.log("POST request on /roles");
+      const newRole = await db.roles.create({
+        role_id: currentId,
+        role_description: req.body.role_description
+      });
+      res.json(newRole);
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  })
+  .put(async (req, res) => {
+    try {
+      console.log("PUT request on /roles");
+      await db.roles.update(
+        {
+          role_description: req.body.role_description
+        },
+        {
+          where: {
+            role_id: req.body.role_id
+          }
+        }
+      );
+      res.send("Successfully Updated");
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  })
+  .delete((req, res) => {
+    res.send("Action unavailable");
+  });
+
+/// /////////////////////////////////
+/// //// Themes Endpoints////////
+/// /////////////////////////////////
+
+router
+  .route("/themes")
+  .get(async (req, res) => {
+    try {
+      console.log("GET request on /themes");
+      const themes = await db.themes.findAll();
+      const reply =
+        themes.length > 0 ? { data: themes } : { message: "no results found" };
+      res.json(reply);
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  })
+  .post(async (req, res) => {
+    const themes = await db.themes.findAll();
+    const currentId = (await themes.length) + 1;
+    try {
+      console.log("POST request on /themes");
+      const newTheme = await db.themes.create({
+        theme_id: currentId,
+        theme: req.body.theme
+      });
+      res.json(newTheme);
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  })
+  .put(async (req, res) => {
+    try {
+      console.log("PUT request on /themes");
+      await db.themes.update(
+        {
+          theme: req.body.theme
+        },
+        {
+          where: {
+            theme_id: req.body.theme_id
+          }
+        }
+      );
+      res.send("Successfully Updated");
+    } catch (err) {
+      console.error(err);
+      res.error("Server error");
+    }
+  })
+  .delete((req, res) => {
+    res.send("Action unavailable");
+  });
+
+/// /////////////////////////////////
+/// //// all_media_backgrounds_link Endpoints////////
+/// /////////////////////////////////
+
+router.route("/mediaBackgroundLinks").get(async (req, res) => {
   try {
-    const halls = await db.DiningHall.findAll();
+    console.log("GET request on /mediaBackgroundLinks");
+    const links = await db.all_media_backgrounds_link.findAll();
     const reply =
-      halls.length > 0 ? { data: halls } : { message: "no results found" };
+      links.length > 0 ? { data: links } : { message: "no results found" };
     res.json(reply);
   } catch (err) {
     console.error(err);
@@ -25,67 +398,17 @@ router.get("/dining", async (req, res) => {
   }
 });
 
-router.get("/dining/:hall_id", async (req, res) => {
-  try {
-    const hall = await db.DiningHall.findAll({
-      where: {
-        hall_id: req.params.hall_id,
-      },
-    });
+/// /////////////////////////////////
+/// //// all_media_creators_link Endpoints////////
+/// /////////////////////////////////
 
-    res.json(hall);
-  } catch (err) {
-    console.error(err);
-    res.error("Server error");
-  }
-});
-
-router.post("/dining", async (req, res) => {
-  const halls = await db.DiningHall.findAll();
-  const currentId = (await halls.length) + 1;
+router.route("/mediaCreatorLinks").get(async (req, res) => {
   try {
-    const newDining = await db.DiningHall.create({
-      hall_id: currentId,
-      hall_name: req.body.hall_name,
-      hall_address: req.body.hall_address,
-      hall_lat: req.body.hall_lat,
-      hall_long: req.body.hall_long,
-    });
-    res.json(newDining);
-  } catch (err) {
-    console.error(err);
-    res.error("Server error");
-  }
-});
-
-router.delete("/dining/:hall_id", async (req, res) => {
-  try {
-    await db.DiningHall.destroy({
-      where: {
-        hall_id: req.params.hall_id,
-      },
-    });
-    res.send("Successfully Deleted");
-  } catch (err) {
-    console.error(err);
-    res.error("Server error");
-  }
-});
-
-router.put("/dining", async (req, res) => {
-  try {
-    await db.DiningHall.update(
-      {
-        hall_name: req.body.hall_name,
-        hall_location: req.body.hall_location,
-      },
-      {
-        where: {
-          hall_id: req.body.hall_id,
-        },
-      }
-    );
-    res.send("Successfully Updated");
+    console.log("GET request on /mediaCreatorLinks");
+    const links = await db.all_media_creators_link.findAll();
+    const reply =
+      links.length > 0 ? { data: links } : { message: "no results found" };
+    res.json(reply);
   } catch (err) {
     console.error(err);
     res.error("Server error");
@@ -93,46 +416,16 @@ router.put("/dining", async (req, res) => {
 });
 
 /// /////////////////////////////////
-/// ////////Meals Endpoints//////////
+/// //// all_media_genres_link Endpoints////////
 /// /////////////////////////////////
-router.get("/meals", async (req, res) => {
-  try {
-    const meals = await db.Meals.findAll();
-    res.json(meals);
-  } catch (err) {
-    console.error(err);
-    res.error("Server error");
-  }
-});
 
-router.get("/meals/:meal_id", async (req, res) => {
+router.route("/mediaGenreLinks").get(async (req, res) => {
   try {
-    const meals = await db.Meals.findAll({
-      where: {
-        meal_id: req.params.meal_id,
-      },
-    });
-    res.json(meals);
-  } catch (err) {
-    console.error(err);
-    res.error("Server error");
-  }
-});
-
-router.put("/meals", async (req, res) => {
-  try {
-    await db.Meals.update(
-      {
-        meal_name: req.body.meal_name,
-        meal_category: req.body.meal_category,
-      },
-      {
-        where: {
-          meal_id: req.body.meal_id,
-        },
-      }
-    );
-    res.send("Meal Successfully Updated");
+    console.log("GET request on /mediaGenreLinks");
+    const links = await db.all_media_genres_link.findAll();
+    const reply =
+      links.length > 0 ? { data: links } : { message: "no results found" };
+    res.json(reply);
   } catch (err) {
     console.error(err);
     res.error("Server error");
@@ -140,54 +433,16 @@ router.put("/meals", async (req, res) => {
 });
 
 /// /////////////////////////////////
-/// ////////Macros Endpoints/////////
+/// //// all_media_themes_link Endpoints////////
 /// /////////////////////////////////
-router.get("/macros", async (req, res) => {
-  try {
-    const macros = await db.Macros.findAll();
-    res.send(macros);
-  } catch (err) {
-    console.error(err);
-    res.error("Server error");
-  }
-});
 
-router.get("/macros/:meal_id", async (req, res) => {
+router.route("/mediaThemeLinks").get(async (req, res) => {
   try {
-    const meals = await db.Macros.findAll({
-      where: {
-        meal_id: req.params.meal_id,
-      },
-    });
-    res.json(meals);
-  } catch (err) {
-    console.error(err);
-    res.error("Server error");
-  }
-});
-
-router.put("/macros", async (req, res) => {
-  try {
-    // N.B. - this is a good example of where to use code validation to confirm objects
-    await db.Macros.update(
-      {
-        meal_name: req.body.meal_name,
-        meal_category: req.body.meal_category,
-        calories: req.body.calories,
-        serving_size: req.body.serving_size,
-        cholesterol: req.body.cholesterol,
-        sodium: req.body.sodium,
-        carbs: req.body.carbs,
-        protein: req.body.protein,
-        fat: req.body.fat,
-      },
-      {
-        where: {
-          meal_id: req.body.meal_id,
-        },
-      }
-    );
-    res.send("Successfully Updated");
+    console.log("GET request on /mediaThemeLinks");
+    const links = await db.all_media_themes_link.findAll();
+    const reply =
+      links.length > 0 ? { data: links } : { message: "no results found" };
+    res.json(reply);
   } catch (err) {
     console.error(err);
     res.error("Server error");
@@ -195,77 +450,16 @@ router.put("/macros", async (req, res) => {
 });
 
 /// /////////////////////////////////
-/// Dietary Restrictions Endpoints///
+/// //// creator_roles_link Endpoints////////
 /// /////////////////////////////////
-router.get("/restrictions", async (req, res) => {
-  try {
-    const restrictions = await db.DietaryRestrictions.findAll();
-    res.json(restrictions);
-  } catch (err) {
-    console.error(err);
-    res.error("Server error");
-  }
-});
 
-router.get("/restrictions/:restriction_id", async (req, res) => {
+router.route("/creatorRoleLinks").get(async (req, res) => {
   try {
-    const restrictions = await db.DietaryRestrictions.findAll({
-      where: {
-        restriction_id: req.params.restriction_id,
-      },
-    });
-    res.json(restrictions);
-  } catch (err) {
-    console.error(err);
-    res.error("Server error");
-  }
-});
-
-/// //////////////////////////////////
-/// ///////Custom SQL Endpoint////////
-/// /////////////////////////////////
-const macrosCustom =
-  "SELECT `Dining_Hall_Tracker`.`Meals`.`meal_id` AS `meal_id`,`Dining_Hall_Tracker`.`Meals`.`meal_name` AS `meal_name`,`Dining_Hall_Tracker`.`Macros`.`calories` AS `calories`,`Dining_Hall_Tracker`.`Macros`.`carbs` AS `carbs`,`Dining_Hall_Tracker`.`Macros`.`sodium` AS `sodium`,`Dining_Hall_Tracker`.`Macros`.`protein` AS `protein`,`Dining_Hall_Tracker`.`Macros`.`fat` AS `fat`,`Dining_Hall_Tracker`.`Macros`.`cholesterol` AS `cholesterol`FROM(`Dining_Hall_Tracker`.`Meals`JOIN `Dining_Hall_Tracker`.`Macros`)WHERE(`Dining_Hall_Tracker`.`Meals`.`meal_id` = `Dining_Hall_Tracker`.`Macros`.`meal_id`)";
-router.get("/table/data", async (req, res) => {
-  try {
-    const result = await db.sequelizeDB.query(macrosCustom, {
-      type: sequelize.QueryTypes.SELECT,
-    });
-    res.json(result);
-  } catch (err) {
-    console.error(err);
-    res.error("Server error");
-  }
-});
-
-const mealMapCustom = `SELECT hall_name,
-  hall_address,
-  hall_lat,
-  hall_long,
-  meal_name
-FROM
-  Meals m
-INNER JOIN Meals_Locations ml 
-  ON m.meal_id = ml.meal_id
-INNER JOIN Dining_Hall d
-ON d.hall_id = ml.hall_id;`;
-router.get("/map/data", async (req, res) => {
-  try {
-    const result = await db.sequelizeDB.query(mealMapCustom, {
-      type: sequelize.QueryTypes.SELECT,
-    });
-    res.json(result);
-  } catch (err) {
-    console.error(err);
-    res.error("Server error");
-  }
-});
-router.get("/custom", async (req, res) => {
-  try {
-    const result = await db.sequelizeDB.query(req.body.query, {
-      type: sequelize.QueryTypes.SELECT,
-    });
-    res.json(result);
+    console.log("GET request on /creatorRoleLinks");
+    const links = await db.creator_roles_link.findAll();
+    const reply =
+      links.length > 0 ? { data: links } : { message: "no results found" };
+    res.json(reply);
   } catch (err) {
     console.error(err);
     res.error("Server error");
